+++ src/Console/Command/LdapSyncCommand.php
@@ -40,7 +40,7 @@ class LdapSyncCommand extends BaseCommand
         $this
             ->addOption('dry-run', null, InputOption::VALUE_NONE)
             ->addOption('create-users', null, InputOption::VALUE_NONE)
-            ->addOption('disable-missed', null, InputOption::VALUE_NONE)
+            ->addOption('disable-missing', null, InputOption::VALUE_NONE)
             ->addOption('no-update', null, InputOption::VALUE_NONE)
             ->addOption('no-disable', null, InputOption::VALUE_NONE)
             ->addOption('no-notify', null, InputOption::VALUE_NONE);
@@ -53,7 +53,7 @@ class LdapSyncCommand extends BaseCommand
         $noUpdate = $input->getOption('no-update');
         $noDisable = $input->getOption('no-disable');
         $noNotify = $input->getOption('no-notify');
-        $disableMissed = $input->getOption('disable-missed');
+        $disableMissing = $input->getOption('disable-missing');
 
         $this->dryrun = $dryrun;
 
@@ -66,7 +66,7 @@ class LdapSyncCommand extends BaseCommand
 
         $this->updateUsers(!$noUpdate);
         $this->disableUsers(!$noDisable);
-        $this->disableMissedUsers($disableMissed);
+        $this->disableMissingUsers($disableMissing);
 
         return 0;
     }
@@ -102,14 +102,14 @@ class LdapSyncCommand extends BaseCommand
     }
 
     /**
-     * Process active users missed from ldap
+     * Process active users missing from ldap
      *
      * @param bool $enabled
      */
-    private function disableMissedUsers($enabled): void
+    private function disableMissingUsers($enabled): void
     {
         if (!$enabled || !$this->ldap->active_dn) {
-            $this->writeln('Skip Disable missed LDAP users', self::VERBOSE);
+            $this->writeln('Skip Disable missing LDAP users', self::VERBOSE);
 
             return;
         }
@@ -122,7 +122,7 @@ class LdapSyncCommand extends BaseCommand
                 // try to find user
                 $remote = $this->ldap->getLdapUser($uid);
 
-                // handle missed users
+                // handle missing users
                 if ($remote === null) {
                    $this->disableAccount("", $uid);
                 }